requirements.txt
python-telegram-bot == 10.1.0
emoji == 0.5.0
pymongo == 3.7.1